Split the weather repository into 2 according to ISP
