+++ app/src/main/java/com/maximapps/maxim_weather/mainScreen/data/repositories/LocationWeatherRepositoryImpl.kt
@@ -2,22 +2,23 @@ package com.maximapps.maxim_weather.mainScreen.data.repositories
 
 import com.maximapps.maxim_weather.mainScreen.data.ResponseMapper
 import com.maximapps.maxim_weather.mainScreen.data.network.WeatherService
-import com.maximapps.maxim_weather.mainScreen.usecases.fetchforecastbyname.CityWeatherRepository
+import com.maximapps.maxim_weather.mainScreen.usecases.fetchforecastbycoordinates.Coordinates
+import com.maximapps.maxim_weather.mainScreen.usecases.fetchforecastbycoordinates.LocationWeatherRepository
 import com.maximapps.maxim_weather.mainScreen.usecases.models.WeatherData
 import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.schedulers.Schedulers
 import javax.inject.Inject
 
 /**
- * Default implementation of the [CityWeatherRepository] interface.
+ * Default implementation of the [LocationWeatherRepository] interface.
  */
-class CityWeatherRepositoryImpl @Inject constructor(
+class LocationWeatherRepositoryImpl @Inject constructor(
     private val service: WeatherService,
     private val mapper: ResponseMapper,
-) : CityWeatherRepository {
+) : LocationWeatherRepository {
 
-    override fun fetchForecastByName(cityName: String): Single<WeatherData> = service
-        .fetchForecast(cityName)
+    override fun fetchForecastByCoordinates(coordinates: Coordinates): Single<WeatherData> = service
+        .fetchForecastByCoordinates(coordinates.latitude, coordinates.longitude)
         .map(mapper::map)
         .subscribeOn(Schedulers.io())
 }
